Implemented game loop, see desc.
The game loop may now be entered through the menu using option 1 for new game. exit command is functional, however save and turn commands are yet to be implemented. While the game loop may be entered, box lid and bag have not, so factories currently print empty at the beginning of the game, also empty spots in the mosaic print as a period rather than a lowercase letter denoting the tile that may be placed there.